Add page-collections
Add a Box and Vec type that are backed by full pages.

These types will be useful when tracking and using storage loaned from
the host in 4k page chunks. Because Salus will be no_std and not have a
local allocator, the smallest unit of memory used will be one page.