riscv-pages: Add CleanPage type
A `CleanPage` represents a page that has been unmapped from a VM and
zeroed. This type will be used as the only way a page can be removed
from a guest by the host. Having this type will ease reasoning about
when pages are cleared and acceptable to map to other VMs.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>