Add fdt module
Add an fdt module that uses fdt-rs to parse the device tree passed from
opensbi.

In addition fdt contains a gross hack that modifies the ram amount
specified to allow the device tree to be reused for the guest.

Both the use of fdt-rs and the hack will be addressed in the future.
fdt-rs has too many dependencies to be used in production.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>