main: Add ability to boot a test VM
This is an initial pass at booting linux as the host VM.

There is now a Makefile that uses qemu to load a kernel at a hard-coded
address, main copies it from that address to the guest's RAM an runs it.

The build process and method of specifying a payload are TBD but will
obviously need to be improved from 'hardcoded to 0c020_0000'.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>