fdt: Extend support for discovering memory regions
Add support for parsing multiple memory regions from the 'memory' node
and for retrieving the reserved memory regions from the FDT header.

Also support looking up the memory regions in which QEMU guest-loader
(via 'multiboot,module' nodes) puts the host VM kernel and initramfs.
We'll use this in a follow-on patch to avoid hard-coding the kernel
load address.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>