Add support for a discontiguous HwMemMap
In order to support discontiguous physical memory and safe loading of the
host VM, add support for having multiple physical memory regions (and
multiple reserved regions) to HwMemMap.

The HwMemMap is built by first inserting the physical memory regions, and
then marking subsets of those regions as reserved. Each reserved region is
associated with a particular HwMemReservedType which allows us to identify
the purpose of the region later on.

Note that this adds a dependency on the arrayvec crate, which itself has
no external dependencies. We use arrayvec here since dynamic memory
allocation isn't available at the point we're constructing the HwMemMap.

This also temporarily limits HypMemPages (and host VM memory) to the first
available contiguous range of system memory. We'll fix this in a follow-on
patch by adding support for discontiguous memory to downstream structures.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>