Implement safe(er) host VM loading
This patch includes a number of changes to page ownership and allocation
at boot time in order to implement safe construction of the host VM's
address space:
  1) Add "free" and "reserved" states to PageInfo. These are used to
     implement a discontiguous PageMap and cleaner ownership semantics:
     pages are free by default until ownership is explicitly taken, first
     by the hypervisor and then, for most pages, by the host.
     Also, use an arrayvec for the per-page owner stack.
  2) Support discontiguous memory maps in PageMap and, by extension,
     HypPageAlloc. This also means that the available host memory returned
     by PageState::from() is now a (possibly-discontiguous) vector of
     page ranges.
  3) Page ownership is now handled exclusively by the higher-level mapping
     logic in {Host,Guest}RootBuilder. Ownership of the page is enforced
     by PlatformPageTable when mapping the page in.

Finally, the host VM building process is reworked such that we build its
address space iteratively, starting at the lowest address. Its kernel and
initramfs images are mapped directly in (rather than copied) and we use
the left over pages returned by PageState::from() to fill in the holes.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>