Add third-party sha2 crate
Measurements of TVM pages are required for attestation, and they will
almost certainly be based on digest algorithms like SHA2. This crate is
a pure Rust implementation of SHA* hashing algorithms, supports
no_std, and is licensed under MIT / Apache 2.0 license.

We'll need an audit to make sure that everything is kosher:
https://gitlab.ba.rivosinc.com/rv/sw/int/salus/-/issues/41