Create a mutable device-tree structure
Add DeviceTree and related structs which allow us to create a mutable
representation of device-tree, either from scratch from an FDT blob.
This is internally implemented as an index-tree, using the Arena type
introduced in the previous patch.

For now, this implementation is meant to cover just the basic structure
of the device-tree (i.e. nodes & properties). It's up to the user to
interpret the properties and to ensure that added properties are
sensible.

We'll use this in a follow-on patch to construct the host VM's device-tree
from the device-tree passed to the hypervisor.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>