Improve host VM device-tree construction
Instead of patching the binary FDT with the new memory node size, use
the new DeviceTree library to build a device-tree for the host VM from
the deserialized hypervisor FDT. This allows us to make much more
significant changes to the device-tree. For now, that's mostly stripping
out all the MMIO devices that the host VM can't access and updating
"chosen" with the location of the initrd.

This change allows Salus to boot Linux all the way to 'init', though it
doesn't do much once it gets there due to the lack of timers.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>