Support serializing DeviceTrees back to an FDT
After we've created a DeviceTree for the host VM, we'll need to turn it
back into an FDT. Add a simple library for serializing a DeviceTree into
an FDT based on v0.3 of the Devicetree Specification.

Note that this is kept in a separate 'serialize' module in order to
isolate from the fdt-rs-dependent cruft.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>