Add CSR type-safety and accessors from Rust
Use the tock_registers crate to generate accessors and register
definitions for the {H,V}S-mode CSRs that we care about. The RISC-V
specifics are borrowed from tock's riscv-csr library with some
modifications: drop support for anything other than RV64, replace
support for M-mode CSRs with their S-mode equivalents, and add
definitions for hypervisor-specific and other S-mode CSRs that were
missing. Also added are helpers for dealing with interrupt/exception
causes and {hg,s}atp values.

This will allow us clean up CSR management (and do it in a type-safe
way), as well as move some of the asm bits to rust.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>