Unify PageRange and SequentialPages
Both are used to represent an owned range of contiguous pages and the
latter has a superset of the functionality of the former, so just turn
all PageRanges into SequentialPages. While we're at it, make the set of
host pages a Vec<> instead of PageVec<> so that we can flatten() it
rather than writing a custom iterator to accomplish the same thing.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>