Run clippy and format in CI
Enable CI to build, clippy, and format that code.

This totally drops our internal CI in favor of a common, but not-audited
external container. This is a temporary hack until we figure out our
internal CI story. It's worth it to carry this ugliness so we can get
some basic checks in as we accelerate development.

Note that tests don't run yet as that needs some more work to get
working with some that cross compile and others that run on the host.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>