Clean up VM enter/exit handling
Many registers are missing from the VM context save/restore and some of
the registers that are included need not be switched out for each VM
(e.g. hideleg, hedeleg). Furthermore, much of the CSR reading and writing
can now safely be done from rust.

First, move the initialization of certain CSRs which need not be touched
when entering or exiting VMs to kernel_init(). Next, divide up VmCpuState
into registers that must be saved/restored on entering/exiting a VM (i.e.
registers that are shared between both worlds), registers that only affect
operation when V=1, and registers which detail the trap cause. Finally,
run_to_exit() and _run_guest are re-written such that only that first
category of registers are swapped in assembly; the rest are handled back
in rust.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>