Stop hard-coding struct offsets in asm
Hard-coding struct offsets in assembly is prone to breakage and human
error. Pass the offsets into the global_asm!() format string instead.
This requires a bit of repitition, but no more so than the C equivalent.

Note that this requires moving the inclusion of guest.S into vm.rs as
asm.rs is also linked in for tellus, which obviously doesn't need these
definitions.

Also move _start to its own linker section to guarantee it appears
first in memory.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>