Allow guests to make use of the Sstc extension
Allow guests to program and take timer interrupts directly at VS level
without hypervisor intervention using the Sstc extension. Since Sstc
support currently requires out-of-tree patches to QEMU, we first need
to check if Sstc is supported by the CPU we're running on using the
'riscv,isa' string in the device-tree. If Sstc is supported, we can
then enable it henvcfg and make sure to save/restore vstimecmp when
context-switching VMs.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>