Remove generic type parameter for measurements
The ergonomic change address the issue of the "bleeding" of generic type parameters,
especially in nested types. In this case, the "" parameter for the measurement type was
causing significant pollution of the type and method signatures, which was hurting
the readability of the code. Furthermore, it necessitated the measurement to support
default instantiation, because the containing types had no real use for it. Also, the
use of the DataMeasure trait was somewhat at cross-purposes with the generic type
parameter; normally, this is needed only to avoid the type erasure associated with
the use of "&dyn" trait, which clearly wasn't the case here. Note that another way
of solving the generic type parameter issue is using enums.

The change also modifies TestMeasure to reflect the above changes.