Add type-safety to supervisor vs guest physical addresses
We're currently using u64 (and sometimes AlignedPageAddr) for guest
physical addresses which can be easily conflated with supervisor
physical addresses. To add some type-safety, turn the PhysAddr type into
RawAddr which takes an AddressSpace as a generic type parameter. The two
types of address spaces we define are SupervisorPhys (for the physical
address space as seen by the hypervisor) and GuestPhys (for the guest
physical address space of VMs). As a result, all types downstream of
RawAddr now are naturally tagged with the address space they belong to.

Update all current users of PhysAddr to use the proper variant for the
type of address they are working with. Also rename AlignedPageAddr to
PageAddr to make the types less verbose now that they're combined with
the address space.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>