Make `PageSize` an enum rather than a trait
Having `PageSize` be a trait means all downstream types must be generic
across page sizes and can make it more difficult to deal with custom
behavior for certain sizes (in paritcular, non-4kB sizes will
eventually require special handling for page ownership). It also
results in having to open-code page table walks, since each page table
level must have its own type.

Make `PageSize` an enum instead and, since 4kB pages are the common
case, make that the default size for `PageAddr` (and by extension, all
downstream types). This then allows us to make `PageTableLevel` a trait
that can be implemented as an enum representing all levels in a paging
structure, which in turn makes `PageTable` and friends generic only
with respect to the top-level `PlatformPageTable` rather than each
individual level. We'll use this in a follow-on patch to simplify
page-table walking.

The downside to all of this of course is that page size and table level
type checks are moved from compile time to run time. Given the
ergonomic improvements it brings to working with page tables (and that
we only care about 4kB pages the vast majority of the time), this seems
worth it, especially when we need to start dealing with huge pages,
MMIO, etc.

Note that any attempts to work with a huge page in `PlatformPageTable`
are now made to error out since ownership for huge pages is totally
unimplemented at the moment.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>