Clean up page table walking
Now that each `PageTable`, `TableEntryMut`, etc does not have a unique
type per level in the paging structure, we can iteratively walk a page
table (e.g. in a loop) vs open-coding the whole walk. Use this to
simplify the various paging operations in `Sv48x4`. Eventually this
could be used to make nearly all of the paging operations generic, and
the individual `PlatformPageTable` implementations would need only
provide the layout.

Note that this also changes the representation of `PageTable` to hold
the address of the page table, rather than a slice of `Pte`s, to allow
for recursing through the page tables without upsetting the borrow
checker. Each `PageTable` is now directly bound to the lifetime of the
root `PlatformPageTable`, rather than indirectly through the parent
`PageTable`.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>