Add a trait representing physical pages of any type
The `Page` structure assumes the backing page is ordinary system RAM,
which makes it not suitable for representing MMIO. Re-introduce the
`PhysPage` trait which can be used to represent a physical page of any
memory type, allowing us to work with the different types of memory
generically while maniupulating page tables. Each implementation of
`PhysPage` must also map to a `MemType`, which we'll use in a
follow-on patch to implement runtime type-checking of pages in
`PageMap`.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>