PlatformPageTable: Support operating on generic page types
Make `PlatformPageTable`'s page table manipulation methods generic
across `PhysPage` implementations. While the caller is responsible for
specifying the type of the page, typing information is cross-checked at
runtime in `PageMap` so that we do not create ordinary RAM `Page`s out
of MMIO (or vice versa).

Note that as a result of this change `unmap_page()` will no longer
clear a page when unmapping it (since it is unsafe to do so for non-RAM
pages). The caller is now responsible for clearing the contents of the
page if it wishes to do so.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>