Add guest VM prototype
This adds a prototype for a guest VM. The image for the guest VM is
concatenated to the Tellus image, and is located using a fixed offset.
The guest VM currently continues to execute until it deliberately
terminates by causing a fault. This change also includes some minor
ergonomic changes to the Makefile.