Forward IMSIC to host VM
We forward the IMSIC to the host VM with guest interrupt file 1 serving
as the host VM's interrupt file and the remaining guest interrupt files
serving as guest interrupt files for the host VM. In other words, we
re-create the IMSIC hierarchy in the guest address space with all
interrupt files shifted down by one page (supervisor -> guest 1,
guest 2 -> guest 1, and so on). In the future the host VM will be able
to delegate its guest interupt files to guest VMs.

Additionally we need to:
 - program hstatus.vgein with the interrupt file ID so that external
   interrupts can be received directly at VS level, and
 - update the host's device tree with the IMSIC node.

With this change, a Linux host VM is able to IPI itself.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>