Add GetSelfMeasurement SBI call
The previous implementation of GetGuestMeasurement overloaded the
semantics of guest_id, and used a guest_id of 0 as a short hand for
self-measurement. While we can continue to support that functionality,
they should be two different APIs per RISC-V SBI conventions. The
change also shuffles the a0...a2 parameters in the GetGuestMeasurement
API for consistency with GetSelfMeasurement.