Make Vm and VmPages thread-safe
In preparation for SMP support, we need to make `Vm`, `VmPages`, and
related structs thread-safe. For `Vm`, introduce separate locks for
vCPU state and guest state, and for `VmPages` introduce separate
locks for the root page table and the measurement. We also need locking
in `GuestRootBuilder` for the vector of page-table pages.

Also, make `PageState`'s methods borrow `self` immutably. We already
have a mutex there for interior mutability.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>