Add basic vCPU management
The `TvmCreate` TEECALL implicitly created the only vCPU up until this
point. In order to support multiple vCPUs, add a `TvmCpuCreate` TEECALL
which takes a vCPU ID and "activates" the vCPU in the TVM. Memory for
vCPUs is still allocated at `TvmCreate` time, but now as a MAX_CPUS-len
`PageVec` of `VmCpu`s. The `TvmCpuRun` TEECALL now respects the
`vcpu_id` argument and will take exclusive ownership of the specified
vCPU while it is running.

Also add a `TvmCpuSetRegister` TEECALL to set vCPU registers. At the
moment it only supports PC and A1, enough to set the entry point for
a VM.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>