Vm: Support Hart State messages from guests
Split the existing "Available" vCPU state into "PoweredOff" and
"Runnable" states, the former being the initial state a vCPU is in
when it is added (other than vCPU0, which we assume to be the boot
CPU).

When a VM requests to start a vCPU, initialize SEPC and A1 as
requested and move the vCPU to the "Runnable" state. This also causes
the trapping vCPU to exit, which we use on the host VM to IPI the
target CPU and have it enter the host VM. While we don't currently
return the exit status for TVM vCPUs, this exit should eventually be
used as a hint that the target vCPU is now runnable.

When a VM requests to stop a vCPU, the vCPU moves itself into the
"PoweredOff" state and exits.

Note that this also requires us to implement more of the Base extension
so that Linux can probe for HSM support.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>