Bring up secondary CPUs
Have the boot CPU allocate a per-CPU region (stack + per-CPU data)
for each CPU present on the system and use the sbi_hart_start SBI
call to boot the secondary CPUs into Salus. For now we just
initialize the secondary CPUs and spin in WFI since we can't run
multiple vCPUs in the host VM yet.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>