Delegate additional exceptions to VMs
We (the hypervisor) can't do much about illegal instructions or
mis-aligned loads/stores, nor would the host VM be able to service
these faults for guest TVMs, so update HEDELG to forward these to
VS level. Because firmware might trap these exceptions first (due
to not being set in MEDELEG), firmware might inject us with the
fault instead of respecting HEDELEG. Handle this case by re-directing
the trap to VS level when we notice that we've exited with an
exception that should have been delegated.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>