Assign VMIDs when entering VMs
Assign VMIDs on a per-CPU basis, fetching a new VMID whenever a vCPU is
first run on a CPU or migrated between CPUs. The per-CPU VMID counter is
versioned in order to handle the rollover of VMIDs. A change in this
version results in a TLB flush and re-assignment of VMIDs for the CPU.

Note that, for now, this scheme results in vCPUs from the same VM using
different VMIDs on the same physical CPU.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>