Track which VM address space is currently active
Rename VmCpu::enter_vm_pages() to VmCpu::activate() and have it return
an ActiveVmCpu, which itself contains an ActiveVmPages that track the
lifetime of our use of the VM's address space. In order to handle VM
nesting, ActiveVmPages restores the previous HGATP when it is dropped.

We'll use ActiveVmPages in a follow-on patch to implement direct access
to guest memory while guaranteeing that the guest's address space is
what's currently active on the running CPU.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>