Copy to/from guest memory using HSV/HLV
Instead of grabbing a lock and walking the page table, use the
hypervisor load/store instructions to make memory accesses in HS-mode
with the active 2nd-stage translation tables.

This requires us to handle guest page faults when using these
instructions, so add a flag to PerCpu that handle_trap() uses to
detect and recover from such faults while we're doing these copies
to/from guest memory.

Now that we're letting the hardware doing the address-checking for us,
we can relax the requirement that the destination address for the
GetMeasurement TEECALLs be page-aligned.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>