Implement TsmGetInfo TEECALL
This allows the host to discover the state of the TSM and how much
memory it needs to donate when creating a TVM. This information is
communicated via a TsmInfo structure that the TSM writes to a
host-specified location in its physical address space.

Use the new TEECALL in Tellus so that we no longer have to hard-code
the number of pages that need to be donated.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>