Make guest VM locking less heavy-handed
All guest VM operations currently require holding the Guests lock in
some form (either shared or exclusive). This makes adding or destroying
a guest a particularly expensive operation as it requires synchronizing
with all operations on all, even completely unrelated, guests.

Improve this situation by making Guests a vector of PageArcs, which
allows us acquire a reference to a guest by cloning the PageArc rather
than holding the Guests lock for the entire lifetime of the reference.

Each GuestState is still protected by a RwLock to support interior
mutability when changing states, however we never block on taking this
lock exclusively since we need an exclusive reference to begin with
when changing states.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>