Decouple size of referenced page from PageAddr
Having PageAddr represent the size of the referenced page is a bit
awkward in the many cases where we use PageAddr to simply represent
an aligned address and do not care about the ownership or size of
the referenced page. Just use PageAddr to represent an address that's
aligned to (at least) the 4kB page size and push PageSize into PhysPage
which actually represents an owned page.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>