Split AddPages into zero-filled and measured variants
Since we'll eventually be unable to map pages from the host into the
child while preserving the contents, we'll need to do a copy from the
source pages to the pages that will get mapped into the guest's address
space when adding non-zero pages. Adding of zero pages can remain the
same. Split the two cases into separate TEECALLs to keep things clean
and avoid optional or overloaded arguments.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>