Add type-safety to Page states
Introduce type states for the various states a Page takes while in
motion within the TSM. The possible states are roughly two-dimensional
describing the usage of the page and its contents.

The "usage" dimension covers:
 - invalidated: pages that have just been invalidated in the page table.
 - converted: pages that have been converted or donated by the owner,
   but not yet assigned to a child VM.
 - internal state: pages that have been assigned to store internal state
   for a VM.
 - mappable: pages that have been assigned to be mapped in a VM.

And the "contents" dimension covers:
 - clean: pages that have been zero-filled.
 - initialized: pages initialized with user-provided data.
 - dirty: pages with unknown or uninitialized contents.

Note that since only a limited combination of these are valid, we keep to
one `State` type.

We then use these types to make page transitions type-safe, e.g. for
conversion or mapping of a page.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>