sbi: Define TVM vCPU exit codes
Define the exit codes to be returned from TvmCpuRun in a TvmCpuExitCode
enum. Since some of these codes require additional information in order
for the host to be able to handle them, we also introduce two virtual
exit cause registers (ExitCause0 and ExitCuase1) to the API and a
TvmCpuGetRegister TEECALL to retrieve the new registers.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>