Introduce PageList for tracking pages in a linked-list
Introduce the PageList container, which is a linked-list of PhysPages
created from linking together PageInfo structs. Since a PhysPage is
a unique reference to a page, a page may only be on at most one list
at a time.

PageList allows us to pass around a dynamically-sized list of pages
without having to allocate storage (e.g. a PageVec) or having to use
an iterator over a page table operation, thereby eliminating the need
to hold the PlatformPageTable lock while copying or clearing pages.
Update invalidate_range() and get_converted_range() to return PageLists
so that they may take advantage of this.

Another type, LockedPageList, is also introduced which is a PageList
specifically for ConvertedLocked pages. Pages that remain on the list
when it is dropped are automatically unlocked, i.e. returned to the
Converted state.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>