PageTable: Move locking inside of PlatformPageTable
Since we're no longer returning iterators with a mutable borrow of
PlatformPageTable, we can make PlatformPageTable handle locking
internally. To facilitate this, make PlatformPageTable a struct rather
than a trait and make the existing implementations zero-sized types
which define the layout of a particular address mode via their
implementation of a PagingMode trait.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>