PageTable: Split page mapping into prepare + map stages
To ease error recovery when mapping a series of pages into a page table,
add a map_pages() API to PlatformPageTable which locks the PTEs and fills
in any intermediate page tables up front, returning a PageTableMapper
object can be used to actually insert the pages and is guaranteed not to
fail as long as the pages are in range (and not repeatedly mapped).

Clean up and improve the type-safety of TableEntryMut while we're at it
by adding type states for Invalid, Leaf, and Table PTEs. Operations on
PTEs (map, invalidate, etc) are now done entirely using TableEntryMut;
PageTable is used only for looking up an entry in a table.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>