Pull page tracking into its own crate
PageTracker, HwMemMap, etc are not page tables nor are they direclty
related to the construction of paging structions, so they are a bit of
an odd fit for the riscv_page_tables crate. Furthermore, we're about to
add a dependency on PageTracker from the various page collection types,
which would create a circular dependency. Move PageTracker and related
types to a page_tracking crate, and move Page{Box,Vec,Arc} into a
collections sub-module of that crate.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>