Page{Box,Arc}: Release backing pages on drop()
Release the backing page of a PageArc or PageBox back to the previous
owner using PageTracker::release_page() when the container is dropped
without having been consumed.

Because PageTracker itself is stored in a leaked PageBox, introduce a
StaticPageRef type that is equivalent to a PageBox that is immediately
leaked.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>