VmPages: Keep track of the guest physical address space map
In preparation for demand-filled pages, shared memory, and emulated MMIO
introduce VmRegionList which keeps track of the regions in the guest
physical address space. For now the only type of region we have is the
memory that's fully populated prior to VM finalization. The region list
is backed by an extra page that the host must donate during TvmCreate.

Note that this could also be accomplished in the page table itself by
flagging PTEs. That wastes much more memory however, in particular for
regions which will never be populated (MMIO) or will likely only be
sparsely populated (shared memory).

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>