Vm: Allow ECALLs to be retried due to a page fault
Some ECALLs require reading from or writing to guest memory. When one
of these calls accesses a valid, but unpopulated, memory region we
need to give the host the ability to resolve the page fault for the
guest. Do this by propagating errors caused by page faults in a guest
ECALL as a ConfidentialPageFault exit to the host and retrying the
faulting ECALL instruction the next time the guest vCPU is run. Note
that page faults in invalid memory regions still cause an error return
to the guest since the fault isn't handleable by the host.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>