sbi: Define TVM emulated MMIO API
Emulated MMIO can be used to support paravirtualized IO, e.g. using the
virtio framework, for TVMs. Define the interfaces for adding emulated MMIO
regions to a TVM and for handling accesses by the TVM to those emulated
MMIO regions.

Emulated MMIO regions always cause a guest page fault and are reported
to the host as a MmioPageFault exit. The faulting address is written
to the virtual ExitCause0 register just like with a ConfidentialPageFault.
The operation the TVM vCPU was attempting to perform (as decoded by the
TSM) is written to the virtual ExitCause1 register, and the operation
itself is completed using the newly-introduced Mmio{Load,Store}Value
virtual registers. For now we only support emulated MMIO operations
corresponding to the RV64I load/store instructions, but it should be
possible to extend this to support emulation of AMO instructions as well.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>