VmPages: Rework guest memory access
Currently we try to fix up any faults in the "in_guest_memcpy" critical
section. Make this more precise by creating an "extable" linker section
that holds the addresses of instructions that are allowed to fault in
HS-mode and only fix up a fault if SEPC matches an entry in the table.
For now this is just the HSV/HLV instructions in copy_{to,from}_guest(),
but, in preparation for fetching instructions from guest memory by guest
virtual address, allow fixing up any exception on matching instructions.

Note that this is very similar to the implementation of the "extable"
facility in Linux for fixing up user access faults, but with some
simplifications.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>