driver: Add PCI configuration space reading
Adds a helper for scanning PCI configuration space.

The core configuration space can return an iterator of the contained
busses and those busses provide device iterators. Devices can provide
iterators across multiple functions if the device supports that.

A top-level function `probe_pci` is added as an initial example. Main
will be able to use it to print the PCI devices that are found(in the
next commit). More advanced usages of PCI will ikely want direct access
to the busses as well as the leaf devices and functions, exposing the
iterators from `config_space` should allow users to build whatever is
needed.

Signed-off-by: Dylan Reid <dgreid@rivosinc.com>