Vm: Handle MMIO page faults
When we take a page fault in an emulated MMIO region, we need to fetch
the faulting instruction in order to communicate the operation to the
VM's host. Add support to ActiveVmPages to fetch the MMIO operation
(including an assembly wrapper for HLVX) and wire up access to the
virtual Mmio{Load,Store}Value registers to support completion of the
MMIO operation.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>