Vm: Trap on WFI in guest VMs
Enable hstatus.VTW so that we trap on WFI in guest VMs and, when a virtual
instruction exception is taken for a WFI in a guest VM, forward the exit
cause to its parent.

Note that we don't trap on WFI for the host VM since all it would do is
add extra overhead.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>