pci: Add type representing a root complex
Add a PcieRoot type which represents a PCIe root complex. For now this
just holds a PciConfigSpace, but will grow to include the other resources
behind the host bridge as well as tracking the busses and devices in the
PCI hierarchy. Also make PcieRoot a singleton to avoid having to pass it
around, and for consistency with the other drivers.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>