pci: Introduce PciFuncConfigSpace to wrap config space accesses
Introduce the PciFuncConfigSpace object which is used to make the raw
MMIO accesses to a PCI function's config space. As part of this change,
Header is made to be a simple container for the attributes read from the
config space header. We'll continue to use Header to enumerate functions
on a bus, but will use PciFuncConfigSpace to implement drivers for the
particular types of functions (bridges or endpoints).

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>