pci: Start building hierarchy of PCI devices
In order to enumerate the PCI hierarchy and support device passthrough
we'll need to maintain some per-device software state. Introduce a few
new types to represent the PCI hierarchy:
 - PciDevice: the common interface for PCI bridges/endpoints
 - PciBus: the set of PCI devices on a bus
 - PciEndpint: an ordinary PCI endpoint
 - PciBridge: a PCI bridge, which itself points to a child bus
PciDevices are maintained in an Arena to simplify the construction and
management of the PCI hierarchy.

Only the root bus is enumerated for now. We'll recursively enumerate
buses behind bridges in a follow-on patch.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>