attestation: Derive the CDI ID and attestation keys from the TSM CDI
We use HMAC-based Key Derivation Function (extract and expand), as
specified in https://datatracker.ietf.org/doc/html/rfc5869, to create a
TVM specific CDI ID and attestation keys to sign the TVM CSRs.

The TVM specific CDI ID is extracted from the TSM one and expanded with
a TVM id.

Then the attestation private key is extracted from this CDI ID, and the
public part is derived from it.

With the attestation manager being able to generate the TVM CDI ID and
attestation key pair, the certificate generation routine now only
depends on an attestation manager instance and the CSR.

Fixes #20

Signed-off-by: Samuel Ortiz <sameo@rivosinc.com>