pci: Support acquiring PhysPages for PCI BAR resource pages
To support mapping PCI BAR resources directly into the host VM's address
space, add a PhysPage type for these pages and an interface to PcieRoot
for taking exclusive ownership over the pages. The configuration space
will be emulated, so we only expose it's location and not an interface
to acquire owning references to the backing pages.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>