pci: Support populating the host device tree
Add support for populating the host's device tree with a PCI node. Most
properties are forwarded as-is under the assumption that the PCI memory
regions will be identity-mapped into the host's address space. Note that
we skip anything related to legacy INTx interrupts since we don't intend
to support their emulation.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>