riscv-regs: Add a trait for CSR access
The ReadWriteRiscvCsr takes the CSR number as a type parameter since CSR
numbers are encoded within the CSR instructions. This makes it impossible
to refer to CSRs generically in, for example, an array. Introduce a
RiscvCsrInterface trait which defines the possible operations that can
be performed on a CSR, which we'll use in a follow-on patch to implement
the array of hpmcounter CSRs.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>