pci: Basic config space emulation
Add support for emulating the standard portion of the PCI configuration
space for each device. We pass through as much control as possible to
the host VM, except:
 - Bus number assignments, since we need requester IDs to remain stable
   for the IOMMU. The bus number assignments are virtualized to the host
   VM and these virtual bus numbers are used when mapping an emulated MMIO
   access to a target device.
 - Legacy PCI bits, expansiom ROMs, and cardbus.
 - INTx, since we don't plan to emulate it.
 - Reset bits.

In addition, we are temporarily not allowing/exposing:
 - Capabilites, since we don't support virtualizing them yet.
 - DMA, since we don't support the IOMMU yet.

Notably, we are allowing the host VM to assign the actual memory and IO
port BAR addresses. This is safe since the host VM is still constrained
to accesses within the BAR resource ranges we map at startup, regardless
of what it programs in the BAR registers.

With this change a Linux host VM is able to enumerate a simple PCI
hierarchy, including assigning BAR resources.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>